backend/requirements.txt
fastapi
uvicorn
celery
redis
requests
beautifulsoup4
pydantic
reportlab
PyJWT
playwright
minio
neo4j
